+++ src/ft_scalar.c
@@ -6,12 +6,12 @@
 /*   By: alfux <alexis.t.fuchs@gmail.com>           +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/29 21:32:39 by alfux             #+#    #+#             */
-/*   Updated: 2022/11/29 21:34:58 by alfux            ###   ########.fr       */
+/*   Updated: 2022/12/29 00:04:50 by alfux            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include <miniRT.h>
 
-float	ft_scalar(t_vec u, t_vec v)
+double	ft_scalar(t_vec u, t_vec v)
 {
 	return (u.x * v.x + u.y * v.y + u.z * v.z);
 }
